requirements.txt
vnstock
google-generativeai
pandas
beautifulsoup4
packaging
requests
